Ran into trouble here with flexible length values such as PrefixedStringVal, whose length can only be known after the string has been read. BinaryReader::fromHandle won't work with PrefixedStringVal for now, the solution will be to add readHandle/writeHandle to the BinVal interface.
However, in real world applications it is probably not a good idea to read
values in increments like one or two bytes, as this will be sloooow - a better
solution would most likely to design the format in a way that larger chunks can
be read into memory at once and then dissected from the string.

The fromString part of BinaryReader should be capable to deal with flexible
length values.