Changed DocBlock type to mixed.
I'm going to redesign all of this. The original intention was to define large
structures using BinStruct, but 'on the field' it turned out to be so overly
complicated that in the End I just resort to IntVal for a small set of values
and StringReader/StringWriter for larger structures.