Replaced IndexedString by String8,16,32,64.
I was never fond of "addIndexedString(8)" in my code. I could have replaced
8 by something like addIndexedString(StringReader::TINY) which was not nice
either, being very long and opening up all sorts of issues how strings with a
32 and 64 bit index should be called. In the end, I find the number to be most
concise, so I went with addString8() and so on. I ommitted Indexed, because in
the end, I never use addCString, so addString8 et al is the default anyway.